Add files via upload
My Project introduces the user to a character named Bones. Bones is attempting to ask the user what their favorite sport might be. The scripts store the name of the user and their favorite sport as items in a list. The script is setup so the user could enter Basketball or basketball, as the condition is searching for a "contains" string of text (in this case "asketball"). After getting his answers, Bones will introduce the user to "Rob Ott", a robot character who will let the user know that he enjoys the same sport. He knows the name and sport of the user by referring to the list items that Bones collected. Rob Ott is also using a loop method that is simulating him rolling back and forth for 185 units of time. After he completes his movement, he is waiting for the user to input the down arrow key, after which he will move to the bottom right corner.